cloud-functions/community-feed-elasticsearch/requirements.txt
functions-framework==3.*
elasticsearch==8.11.0
google-generativeai==0.3.2
